pyproject.toml
[project]
name = "autonomous-hacks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.10.0",
    "fastapi>=0.125.0",
    "fpdf2>=2.8.5",
    "google-genai>=1.56.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "aiohttp>=3.9.0",
    "ics>=0.7.2",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.45",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "argon2-cffi>=23.1.0",
    "streamlit>=1.52.2",
    "requests>=2.32.5",
    "watchdog>=6.0.0",
    "ty>=0.0.8",
]

[dependency-groups]
dev = [
    "bandit>=1.9.2",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.14.10",
]
